requirements.txt
# requirements.txt
# Standard libraries for data handling and computation
numpy==1.26.4
pandas==1.3.5
scipy==1.14.1

# Machine Learning and NLP libraries
torch==2.4.0+cu116  # Specify CUDA version if needed
torchvision==0.19.0+cu116  # Specify CUDA version if needed
transformers==4.15.0
sentence-transformers==2.2.0

# For efficient nearest neighbor search
faiss-gpu==1.8.0  # Make sure to use the GPU version for CUDA compatibility

# Utilities
tqdm==4.62.3
setuptools_rust  # Required for compiling tokenizers with Rust dependencies
tokenizers==0.10.3  # Ensure compatibility with transformers

# If you're using Rust-based libraries
rust==2024.1  # Pseudo-version, adjust based on your need or omit if not using Rust directly

# Additional Python packages that may be required
huggingface-hub==0.24.6
